Add all 25 reverse coercion functions of the form "floattoint", "doubletochar"... Still need to do stringtochar and chartostring.
Documentation will be forthcoming

-ethan


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@3177 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddIntrinsics.c
@@ -1,7 +1,7 @@
 
 
 /*
- *      $Id: AddIntrinsics.c,v 1.14 1995-04-12 00:04:46 ethan Exp $
+ *      $Id: AddIntrinsics.c,v 1.15 1995-04-14 22:02:05 ethan Exp $
  */
 /************************************************************************
 *									*
@@ -539,7 +539,7 @@ void _NclAddIntrinsics
 	_NclRegisterFunc(_NclIgetenv,args,"getenv",1,IFUNC);
 
 	args = NclCalloc(1,sizeof(NclArgTemplate));
-	args[0].arg_data_type = _NclLookUp("numeric");
+	args[0].arg_data_type = _NclLookUp("integer");
 	args[0].is_dimsizes = 0;
 	_NclRegisterFunc(_NclIabs,args,"abs",1,IFUNC);
